src/bleu_type_metrics/requirement.txt
numpy
tqdm
torch
torchvision
gensim
mecab-python3
sentencepiece
scipy
nltk
scikit-learn
